# improved maps:with and maps:without algorithm