Inspired by the analogous elixir implementations, this PR improves the performance of `maps:with` and `maps:without`, particularly in the case of large maps.

The current implementation is roughly O(N*M) where N is the number of items to be removed, and M is the number of items in the map. This does not include the cost of `maps:from_list` or `maps:to_list`. This leads to pretty horrifying execution times on large maps.

The new implementation is O(N) where N is the number of items to be removed. For each N there's the cost of removing a key from a map, and while I don't know that off hand, in practice that turns out to be a vast improvement for all map sizes I tested.

Examples:
Map size: 1,000,000. Keys to remove: 1,000.
Old: 4,312.239 ms
New: 3.141 ms.

Map size: 1,000. Keys to remove: 100.
Old: 9.435 ms
New: 0.077 ms.