@essen 2016-02-09
Curious as to what happens under the hood though. The old code has Map1, creates list L and then creates Map2 back. Doesn't the new code create a lot of intermediary maps, 1 for each key added/removed, which would have a cost possibly paid off at the next GC?

It would be interesting to see if the results change by adding a GC of the process just before the end of the benchmark (and have one just before starting it to make sure there's no interferences).

You might also want to post the code you used to benchmark also.