@ericmj 2016-02-09
Another approach to maps:with/2 would be to fold into a list and create a single map. It would still iterate over the keys to take instead of over the whole map. Don't know how maps:from_list/1 works under the hood so can't say for sure if it will be more performant.