@psyeugenic 2016-02-09
The original `with/2` and `without/2` code could certainly be improved. What you are doing here is a trade-off, more speed with more gc, normally not a problem.

Under the hood, `maps:from_list/1` is a beast. It has to be a BIF (or at least the equivalent needs to be accessible to the runtime. I haven't gone back to improve the code since I originally wrote it. I'm confident it could be improved. See, https://github.com/erlang/otp/blob/maint/erts/emulator/beam/erl_map.c#L504-L899

The `maps:to_list/1` creates unnecessary garbage. What we lack is an iterator for maps.