@kostis 2016-02-09
W.r.t. improving the code of `with/2` and `without/2` a simple improvement is the following.  Instead of
<pre>maps:from_list([{K,V}||{K,V} <- maps:to_list(M), not lists:member(K, Ks)]);</pre>
write:
<pre>maps:from_list([KV || KV={K,V} <- maps:to_list(M), not lists:member(K, Ks)]);</pre>
which avoids the (completely unnecessary) construction of a new pair.