@benwilson512 2016-02-09
Thanks for the feedback folks.

@ericmj great point on `maps:take`, local testing seems to indicate that building a [{K, V}] list and calling `maps:from_list` on it does better than N `maps:put` calls.

@psyeugenic would you mind elaborating on the way in which the new approach generates more work for the GC ? On the one hand it creates N copies of the map to be GCed. On the other hand, the current approach seems to create a list M items long that would presumably also need GCed.

@kostis unfortunately this still has the same O(N*M) issue as the original algorithm.