@fenollp 2016-02-09
@kostis Are you confident this avoids the construction of a new pair? I thought (hoped) that the compiler would take care of this itself.