@psyeugenic 2016-02-09
@benwilson512 Just that it might increase the number of gc's and if you have a large live heap it could be costly. But I wouldn't worry about it too much. Your suggestion is better algorithmically anyways.

Both `maps:from_list/1` and iterating over all keys and doing `maps:put/3` are O(N * log N) dominated but `maps:from_list/1` has lower constant in front of that O. (from_list does a quicksort on hashvalues which makes it O(N * log N)).