@benwilson512 2016-02-09
Excellent. I pushed a small change based on @ericmj's suggestion for `maps:take`. What's the process from here?