@psyeugenic 2016-02-09
Now is the waiting phase ^^

This has to churn through our CI machinery and visit mandatory testing.

I will nitpick your commit message though. Squash the two commits and use an imperative voice in the message, in this case I would use "Improve maps:with/2 and maps:without/2 algorithms". You could also include why the change was done, like you did above. I.e. Squash, edit and force-push your branch.

See, https://github.com/erlang/otp/wiki/Writing-good-commit-messages

If nothing is amiss it will be merged, otherwise we will get back to you. But fix the commits.